- Apparently forgot to update RP commands on GitHub, so all RP commands now have cooldown - Filter now has support for numbers in the filter list - Removed a few unused constants for code cleanup

+++ commands/roleplay/dropkick.js
@@ -3,11 +3,17 @@ module.exports = {
     usage: "< id / mention >",
     run: async function (client, message, args) {
 
-        if(message.deletable) message.delete();
+        if (client.cooldownManager.checkCooldownAndNotify("rp", message.author.id, message)) {
+            return;
+        }
+
+        if (message.deletable) message.delete();
 
         let dkUser = message.mentions.members.first() || message.guild.members.cache.get(args[0]);
         if (!dkUser) return message.reply("user not found.");
         if (message.mentions.members.first().user === message.author) return message.reply("you can't roleplay with yourself!");
+
+        client.cooldownManager.setCoolDown("rp", message.author.id, 60);
 
         await message.channel.send(`${message.author} has drop-kicked ${dkUser} all the way to the next country!`);
     }
